// types.ts
export interface IRecords {
    'directory_category' : string;
    'content_children_count': string;
    'directory_contact__email' : string;
    'directory_contact__fax' : string;
    'directory_contact__mobile' : string;
    'directory_contact__phone': string;
    'directory_contact__website': string;
    'content_post_id': string;
    'content_post_slug': string;
    'content_body': string;
    'directory_location__street': string;
    'directory_location__city': string;
    'directory_location__country': string;
    'directory_location__address': string;
    'directory_location__lat': string;
    'directory_location__lng': string;
    'directory_location__zip': string;
    'directory_location__state': string;
    'directory_social__facebook': string;
    'directory_social__googleplus': string;
    'directory_social__twitter': string;
    'content_post_status': string;
    'content_post_title': string;
}